Create palindrome2
This is an algorithm from Euler Project. This algorithm allows you to find the largest palindromic number possible when multiplying two numbers together with a certain number of digits. For instance, if you pass the function palin() with 3 as an argument, it will give you the largest palindromic number possible from multiplying two three digit numbers together. This algorithm returns an array with the first number being the largest palindromic number possible, with the next two being the multiplicand and multiplier.